create command factory, workflow runner refactor
