--- tt-inference-server-v2/empty.py
@@ -1,3 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-#
-# SPDX-FileCopyrightText: © 2026 Tenstorrent USA, Inc.
