+++ tt-inference-server-v2/workflow_module/__init__.py
@@ -20,6 +20,16 @@ _LAZY_FROM_WORKFLOWS = {
     "get_workflow_class",
 }
 
+_LAZY_FROM_COMMANDS = {
+    "Command",
+    "CommandResult",
+    "WorkflowCommand",
+}
+
+_LAZY_FROM_COMMAND_FACTORY = {
+    "CommandFactory",
+}
+
 
 def __getattr__(name):
     if name in _LAZY_FROM_EXECUTION:
@@ -30,6 +40,14 @@ def __getattr__(name):
         from . import workflows
 
         return getattr(workflows, name)
+    if name in _LAZY_FROM_COMMANDS:
+        from . import commands
+
+        return getattr(commands, name)
+    if name in _LAZY_FROM_COMMAND_FACTORY:
+        from . import command_factory
+
+        return getattr(command_factory, name)
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
 
 
@@ -39,4 +57,6 @@ __all__ = [
     "get_default_accumulator",
     *sorted(_LAZY_FROM_EXECUTION),
     *sorted(_LAZY_FROM_WORKFLOWS),
+    *sorted(_LAZY_FROM_COMMANDS),
+    *sorted(_LAZY_FROM_COMMAND_FACTORY),
 ]
