+++ tt-inference-server-v2/workflow_module/commands.py
@@ -1,0 +1,90 @@
+# SPDX-License-Identifier: Apache-2.0
+#
+# SPDX-FileCopyrightText: 2026 Tenstorrent AI ULC
+
+from __future__ import annotations
+
+import logging
+from abc import ABC, abstractmethod
+from dataclasses import dataclass
+from typing import Any, Optional
+
+from test_module import MediaContext
+
+from .execution import OrchestratorMetadata, WorkflowResult
+
+logger = logging.getLogger(__name__)
+
+
+@dataclass(frozen=True)
+class CommandResult:
+    command_name: str
+    return_code: int
+    error: Optional[str] = None
+    payload: Optional[Any] = None
+
+    @property
+    def succeeded(self) -> bool:
+        return self.return_code == 0
+
+
+class Command(ABC):
+    name: str = ""
+
+    @abstractmethod
+    def execute(self) -> CommandResult: ...
+
+
+class WorkflowCommand(Command):
+    name = "workflow"
+
+    def __init__(
+        self,
+        ctx: MediaContext,
+        *,
+        workflow_name: str,
+        orchestrator_metadata: OrchestratorMetadata,
+        num_prompts: Optional[int] = None,
+    ) -> None:
+        self.ctx = ctx
+        self.workflow_name = workflow_name
+        self.orchestrator_metadata = orchestrator_metadata
+        self.num_prompts = num_prompts
+
+    def execute(self) -> CommandResult:
+        from .blocks_sink import get_default_accumulator
+        from .workflows import get_workflow_class
+
+        self._apply_num_prompts_override()
+        get_default_accumulator().clear()
+        workflow_cls = get_workflow_class(self.workflow_name)
+        workflow = workflow_cls(
+            self.ctx,
+            orchestrator_metadata=self.orchestrator_metadata,
+        )
+        result: WorkflowResult = workflow.run()
+        return CommandResult(
+            command_name=self.name,
+            return_code=result.return_code,
+            error=result.error,
+            payload=result,
+        )
+
+    def _apply_num_prompts_override(self) -> None:
+        if self.num_prompts is None:
+            return
+        from test_module.benchmark_tests import image_benchmark_tests as _ibt
+
+        _ibt.SDXL_BENCHMARK_NUM_PROMPTS = self.num_prompts
+        _ibt.SDXL_SD35_BENCHMARK_NUM_PROMPTS = self.num_prompts
+        logger.info(
+            "Overriding image benchmark + spec_tests prompt count to %d",
+            self.num_prompts,
+        )
+
+
+__all__ = [
+    "Command",
+    "CommandResult",
+    "WorkflowCommand",
+]
